♻ Move color utils into util module
